Fix setting session expiry when using memcached for sessions.
This commit fixes a bug where the ini file is incorrectly parsed resulting in the session ttl always being set to 0.
This leads to never expiring sessions, bloating memcached usage. This is particularly troublesome when using
CouchBase as a memcached replacement as sessions never expire, consuming all CouchBase storage.

Markus Berthold found and fixed the bug. For more details, see:
http://www.couchbase.org/forums/thread/using-membase-php-session-handler