Fixed broken test tests/cas.phpt.
Now, $cas_tokens doesn't need to be passed by reference in the get, getMulti, getByKey, getMultiByKey.
Both by reference and by value are supported, but cas will only be enabled if it's passed by reference.
It's handy if you want to use the udf_flags but you don't want to enable cas.