First pass at using simple memcached_get when possible
In binary mode, memcached_mget always issues two packets, which ends up
making it slower than ascii mode for many users.

This commit incorrectly believes that libmemcached supports UDP get, it
does not, so that assumption needs to be reworked. Maybe UDP mode should
automatically make both a TCP and a UDP connection? Or return a class
with fewer methods so that it can be introspected or used with a magic
method not defined to fall back to a TCP connection class as needed?